+++ src/components/GlobalErrorBoundary/index.tsx
@@ -3,7 +3,7 @@ import { Text, Link, Icon, FixedIcon, Title } from '@gnosis.pm/safe-react-compon
 import { IS_PRODUCTION } from 'src/utils/constants'
 import { FallbackRender } from '@sentry/react/dist/errorboundary'
 import { ROOT_ROUTE } from 'src/routes/routes'
-import { loadFromSessionStorage, removeFromSessionStorage, saveToSessionStorage } from 'src/utils/storage'
+import { loadFromSessionStorage, removeFromSessionStorage, saveToSessionStorage } from 'src/utils/storage/session'
 
 const Wrapper = styled.div`
   width: 100%;
@@ -56,8 +56,10 @@ export const handleChunkError = (error: Error): boolean => {
 
   const lastReload = loadFromSessionStorage<number>(LAST_CHUNK_FAILURE_RELOAD_KEY)
 
+  const isTimestamp = typeof lastReload === 'number' && !isNaN(lastReload)
+
   // Not a number in the sessionStorage
-  if (typeof lastReload !== 'number' || isNaN(lastReload)) {
+  if (!isTimestamp) {
     removeFromSessionStorage(LAST_CHUNK_FAILURE_RELOAD_KEY)
     return false
   }
